+++ languagetool-server/src/test/java/org/languagetool/server/TextCheckerTest.java
@@ -26,10 +26,7 @@ import org.junit.Test;
 import org.languagetool.markup.AnnotatedTextBuilder;
 
 import java.io.*;
-import java.util.Arrays;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.*;
@@ -38,6 +35,7 @@ public class TextCheckerTest {
 
   private final String english = "This is clearly an English text, should be easy to detect.";
   private final TextChecker checker = new V2TextChecker(new HTTPServerConfig(), false, null, new RequestCounter());
+  private final String unsupportedCzech = "V současné době je označením Linux míněno nejen jádro operačního systému, ale zahrnuje do něj též veškeré programové vybavení";
 
   @Test
   public void testMaxTextLength() throws Exception {
@@ -128,26 +126,40 @@ public class TextCheckerTest {
 
   @Test
   public void testDetectLanguageOfString() {
-    assertThat(checker.detectLanguageOfString("", "en", Arrays.asList("en-GB")).getShortCodeWithCountryAndVariant(), is("en-GB"));
-    assertThat(checker.detectLanguageOfString("X", "en", Arrays.asList("en-GB")).getShortCodeWithCountryAndVariant(), is("en-GB"));
-    assertThat(checker.detectLanguageOfString("X", "en", Arrays.asList("en-ZA")).getShortCodeWithCountryAndVariant(), is("en-ZA"));
-    assertThat(checker.detectLanguageOfString(english, "de", Arrays.asList("en-GB", "de-AT")).getShortCodeWithCountryAndVariant(), is("en-GB"));
-    assertThat(checker.detectLanguageOfString(english, "de", Arrays.asList()).getShortCodeWithCountryAndVariant(), is("en-US"));
-    assertThat(checker.detectLanguageOfString(english, "de", Arrays.asList("de-AT", "en-ZA")).getShortCodeWithCountryAndVariant(), is("en-ZA"));
+    List<String> e = Collections.emptyList();
+    assertThat(checker.detectLanguageOfString("", "en", Arrays.asList("en-GB"), e).getShortCodeWithCountryAndVariant(), is("en-GB"));
+    assertThat(checker.detectLanguageOfString("X", "en", Arrays.asList("en-GB"), e).getShortCodeWithCountryAndVariant(), is("en-GB"));
+    assertThat(checker.detectLanguageOfString("X", "en", Arrays.asList("en-ZA"), e).getShortCodeWithCountryAndVariant(), is("en-ZA"));
+    assertThat(checker.detectLanguageOfString(english, "de", Arrays.asList("en-GB", "de-AT"), e).getShortCodeWithCountryAndVariant(), is("en-GB"));
+    assertThat(checker.detectLanguageOfString(english, "de", Arrays.asList(), e).getShortCodeWithCountryAndVariant(), is("en-US"));
+    assertThat(checker.detectLanguageOfString(english, "de", Arrays.asList("de-AT", "en-ZA"), e).getShortCodeWithCountryAndVariant(), is("en-ZA"));
     String german = "Das hier ist klar ein deutscher Text, sollte gut zu erkennen sein.";
-    assertThat(checker.detectLanguageOfString(german, "fr", Arrays.asList("de-AT", "en-ZA")).getShortCodeWithCountryAndVariant(), is("de-AT"));
-    assertThat(checker.detectLanguageOfString(german, "fr", Arrays.asList("de-at", "en-ZA")).getShortCodeWithCountryAndVariant(), is("de-AT"));
-    assertThat(checker.detectLanguageOfString(german, "fr", Arrays.asList()).getShortCodeWithCountryAndVariant(), is("de-DE"));
+    assertThat(checker.detectLanguageOfString(german, "fr", Arrays.asList("de-AT", "en-ZA"), e).getShortCodeWithCountryAndVariant(), is("de-AT"));
+    assertThat(checker.detectLanguageOfString(german, "fr", Arrays.asList("de-at", "en-ZA"), e).getShortCodeWithCountryAndVariant(), is("de-AT"));
+    assertThat(checker.detectLanguageOfString(german, "fr", Arrays.asList(), e).getShortCodeWithCountryAndVariant(), is("de-DE"));
+    assertThat(checker.detectLanguageOfString(unsupportedCzech, "en", Arrays.asList(), e).
+            getShortCodeWithCountryAndVariant(), is("sk-SK"));  // misdetected because it's not supported
+  }
+
+  @Test
+  @Ignore("requires fastText (binary and model) installed locally")
+  public void testDetectLanguageOfStringWithFastText() {
+    HTTPServerConfig config = new HTTPServerConfig();
+    config.setFasttextBinary(new File("/prg/fastText-0.1.0/fasttext"));
+    config.setFasttextModel(new File("/prg/fastText-0.1.0/data/lid.176.bin"));
+    TextChecker checker = new V2TextChecker(config, false, null, new RequestCounter());
+    assertThat(checker.detectLanguageOfString(unsupportedCzech, "en", Arrays.asList(), Arrays.asList("foo", "cs")).
+            getShortCodeWithCountryAndVariant(), is("zz"));  // cs not supported but mapped to noop language 
   }
 
   @Test(expected = RuntimeException.class)
   public void testInvalidPreferredVariant() {
-    checker.detectLanguageOfString(english, "de", Arrays.asList("en"));  // that's not a variant
+    checker.detectLanguageOfString(english, "de", Arrays.asList("en"), Collections.emptyList());  // that's not a variant
   }
 
   @Test(expected = RuntimeException.class)
   public void testInvalidPreferredVariant2() {
-    checker.detectLanguageOfString(english, "de", Arrays.asList("en-YY"));  // variant doesn't exist
+    checker.detectLanguageOfString(english, "de", Arrays.asList("en-YY"), Collections.emptyList());  // variant doesn't exist
   }
 
 }